+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change history for ui-dashboard
 
-## 4.1.0 In progress
+## 5.0.0 In progress
+  * STRIPES-870 BREAKING upgrade react to v18
+    * ERM-2996 Upgrade ui-dashboard React to v18
 
 ## 4.0.0 2023-02-22
   * ERM-2575 fix deprecated default export from 'zustand' in ui-dashboard